requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0

# Database
# sqlite3 is built-in with Python, no installation needed

# Configuration
python-dotenv==1.0.0

# Utilities
urllib3==2.1.0

# Future SMS integration (commented out for now)
# twilio==8.11.1
